Pull learners, features, and datasets into smlb/ directory. Fix test imports
