+++ smlb/datasets/experimental/band_gaps_sc73/band_gap_jpcrd_transformed-pif-onlyRef.json.zip
