+++ smlb/datasets/synthetic/friedman_1979/friedman_1979.py
