+++ smlb/datasets/synthetic/friedman_silverman_1989/friedman_silverman_1989.py
