+++ smlb/datasets/synthetic/morse_potential/morse_potential.py
