+++ tests/test_workflows.py
@@ -7,12 +7,10 @@ A benchmark of regression models in chem- and materials informatics.
 
 import pytest
 
-import numpy as np
-
 pytest.importorskip("sklearn")
 
 import smlb
-from smlb import params
+
 
 #############################
 #  LearningCurveRegression  #
@@ -22,7 +20,7 @@ from smlb import params
 def test_learning_curve_regression():
     """Simple examples"""
 
-    from datasets.synthetic.friedman_1979.friedman_1979 import Friedman1979Data
+    from smlb.datasets.synthetic.friedman_1979.friedman_1979 import Friedman1979Data
 
     dataset = Friedman1979Data(dimensions=5)
 
@@ -32,14 +30,14 @@ def test_learning_curve_regression():
         smlb.RandomVectorSampler(size=size, rng=0) for size in training_sizes
     )  # dataset domain is used by default
 
-    from learners.scikit_learn.gaussian_process_regression_sklearn import (
+    from smlb.learners.scikit_learn.gaussian_process_regression_sklearn import (
         GaussianProcessRegressionSklearn,
     )
 
     learner_gpr_skl = GaussianProcessRegressionSklearn(
         random_state=0
     )  # default is Gaussian kernel
-    from learners.scikit_learn.random_forest_regression_sklearn import (
+    from smlb.learners.scikit_learn.random_forest_regression_sklearn import (
         RandomForestRegressionSklearn,
     )
